+++ dbms/src/Interpreters/ProcessList.h
@@ -65,7 +65,7 @@ struct QueryStatusInfo
     bool is_cancelled;
 
     /// Optional fields, filled by query
-    std::vector<UInt32> thread_ids;
+    std::vector<UInt64> thread_ids;
     std::shared_ptr<ProfileEvents::Counters> profile_counters;
     std::shared_ptr<Settings> query_settings;
 };
