+++ dbms/src/Storages/MergeTree/MergeList.h
@@ -45,7 +45,7 @@ struct MergeInfo
     UInt64 rows_written;
     UInt64 columns_written;
     UInt64 memory_usage;
-    UInt32 thread_id;
+    UInt64 thread_id;
 };
 
 struct FutureMergedMutatedPart;
@@ -87,7 +87,7 @@ struct MergeListElement : boost::noncopyable
     MemoryTracker * background_thread_memory_tracker;
     MemoryTracker * background_thread_memory_tracker_prev_parent = nullptr;
 
-    UInt32 thread_id;
+    UInt64 thread_id;
 
 
     MergeListElement(const std::string & database, const std::string & table, const FutureMergedMutatedPart & future_part);
